% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_curves.R
\name{get_rarefy_curves}
\alias{get_rarefy_curves}
\title{Calculate the rarefaction curves}
\usage{
get_rarefy_curves(
  asm,
  n_perm = 999,
  n_points = 100,
  max_indiv = 10000,
  step_scale = c("log", "sqrt", "linear"),
  probs = c(0.1, 0.5, 0.9),
  indices = c("CWM", "sumT", "rangeT", "S"),
  null_models = c("emp", "null_Tr", "null_SAD", "null_all", "flat")
)
}
\arguments{
\item{asm}{a \code{trasm} data object}

\item{n_perm}{number of permutations per null model}

\item{n_points}{number of estimates per curve (needed?)}

\item{max_indiv}{maximum individuals sampled for the rarefaction curves}

\item{step_scale}{scale of the steps to build the rarefaction curves}

\item{probs}{quantile of the permutation probabilities}

\item{indices}{trait indices to be calculated per null model}

\item{null_models}{null models to be calculated}
}
\value{
A \code{data.frame} with community attributes
}
\description{
from assemblages in long format
calculate the rarefaction curves based on different null models
...
to be completed
}
\examples{
asm <- simulate_assemblage(S = 20, N = 1000)
cur <- get_rarefy_curves(asm, n_perm = 100, n_points = 100)
gplot_curves(cur, indices = c("S", "CWM"), null_models = c("emp", "flat"))

}
