% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gplot_curves}
\alias{gplot_curves}
\title{Plot the rarefaction curves}
\usage{
gplot_curves(
  curves,
  indices = c("CWM", "sumT", "rangeT", "S"),
  null_models = c("emp", "null_Tr", "null_SAD", "null_all", "flat"),
  title = "Accumulation curves",
  show_ci = TRUE
)
}
\arguments{
\item{curves}{a \code{data.frame} object output of rarefy_curves}

\item{indices}{number of permutations per null model}

\item{null_models}{number of estimates per curve (needed?)}

\item{title}{maximum individuals sampled for the rarefaction curves}

\item{show_ci}{logical argument to show confidence interval}
}
\value{
A \code{ggplot2} object
}
\description{
plot the metrics and the different
calculate the rarefaction curves based on different null models
...
to be completed
}
\examples{
asm <- simulate_assemblage(S = 20, N = 1000)
cur <- get_rarefy_curves(asm, n_perm = 100, n_points = 100)
gplot_curves(cur)

}
