% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gplot_effects}
\alias{gplot_effects}
\title{Plot the effect curves}
\usage{
gplot_effects(
  curves,
  indices = c("CWM", "sumT", "rangeT", "S"),
  effects = c("covTA", "SAD", "traits", "SADflat"),
  show_ci = TRUE,
  palette = "Set1"
)
}
\arguments{
\item{curves}{a \code{data.frame} object output of rarefy_curves}

\item{indices}{number of permutations per null model}

\item{effects}{number of estimates per curve (needed?)}

\item{show_ci}{logical argument to show confidence interval}

\item{palette}{color palette}
}
\value{
A \code{ggplot2} object
}
\description{
plot the effect curves
...
to be completed
}
\examples{
asm <- simulate_assemblage(S = 20, N = 1000)
cur <- get_rarefy_curves(asm, n_perm = 100, n_points = 100)
gplot_effects(cur)

}
