% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasm.R
\name{new_trasm}
\alias{new_trasm}
\title{Create the 'trasm' object.}
\usage{
new_trasm(asm, species_var = "Sp", trait_var = "Tr", abundance_var = "Ab")
}
\arguments{
\item{asm}{data.frame in long format with columns containing species, traits, and abundance information.}

\item{species_var}{column name containing species information}

\item{trait_var}{column name containing trait information}

\item{abundance_var}{column name containing abundance information}
}
\value{
a "trasm" object which is a \code{data.frame} with three columns: Sp, Tr, and Ab
}
\description{
The 'trasm' object is a trait assemblage dataset in long format
that allows the calculation of trait-abundance scaling relationship across scales.
}
