% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition the effects from the rarefaction curves}
\usage{
partition(
  curves,
  conservative_ci = TRUE,
  add_percent = TRUE,
  eps = 1e-08,
  long_format = FALSE
)
}
\arguments{
\item{curves}{a \code{trasmout} data object}

\item{conservative_ci}{calculate the confidence intervals}

\item{add_percent}{calculate percentages}

\item{eps}{threshold value}

\item{long_format}{return long formatted data.frame}
}
\value{
A \code{data.frame} with partition effects in long or wide format
}
\description{
partition the effect between covTA, SAD, traits, SAD_flat
...
to be completed
}
\examples{
asm <- simulate_assemblage(S = 20, N = 1000)
cur <- get_rarefy_curves(asm, n_perm = 100, n_points = 100)
partition(cur)

}
