% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_scale}
\alias{seq_scale}
\title{Returns a uniformly distributed steps along scale}
\usage{
seq_scale(
  min_indiv = 1,
  max_indiv = 1000,
  n_points = 10,
  step_scale = c("log", "sqrt", "linear")
)
}
\arguments{
\item{min_indiv}{minimum number of}

\item{max_indiv}{maximum number of}

\item{n_points}{number of points estimated in the curves}

\item{step_scale}{scale of the steps to calculate the curves (not clear)}
}
\value{
A \code{data.frame} with estimated trait-abundance relationship
}
\description{
Make a monotonically increasing integer vector of steps in [min_indiv, max_indiv],
with denser spacing near small m when mode = "log" or "sqrt".
}
\examples{
plot(seq_scale(1, 1000, 10, "log"), type="l")
lines(seq_scale(1, 1000, 10, "sqrt"), col="red")
lines(seq_scale(1, 1000, 10, "linear"), col="blue")

}
