% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_assemblage}
\alias{simulate_assemblage}
\title{Clean and homogenized city names}
\usage{
simulate_assemblage(
  S = 20,
  N = 1000,
  trait_mean = 1,
  trait_sd = 1,
  sad_k = 1,
  beta_cov = 0.8
)
}
\arguments{
\item{S}{number of species}

\item{N}{number of individuals}

\item{trait_mean}{mean value of the trait}

\item{trait_sd}{standard deviation of the trait}

\item{sad_k}{dispersion of the neagitve binomial, closer to 0 means more unequal SAD}

\item{beta_cov}{standard deviation of the trait}
}
\value{
A \code{data.frame} with community attributes
}
\description{
Create random communities with:
\itemize{
\item normally distributed trait values
\item negatively distributed individuals among species
\item trait abundance coupling
}
}
\examples{
simulate_assemblage(S=20, N=1000)

}
