% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasm.R
\name{summary.trasm}
\alias{summary.trasm}
\title{Calculate a short summary of the trasm data}
\usage{
\method{summary}{trasm}(object, ...)
}
\arguments{
\item{object}{\code{trasm} object}

\item{...}{left over}
}
\description{
Calculate a short summary of the trasm data
}
