% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasm.R
\name{summary_trasm}
\alias{summary_trasm}
\title{Calculate a short summary of the trasm data}
\usage{
summary_trasm(x)
}
\arguments{
\item{x}{\code{trasm} object}
}
\description{
Calculate a short summary of the trasm data
}
